ATS_CV_Optimizer_Pro_FINAL/requirements.txt

requests

python-docx
